requirements.txt
Flask==2.2.5
flask-cors==5.0.1
numpy==1.23.5
pillow==10.2.0